requirements.txt
streamlit>=1.40.0
databricks-sdk>=0.35.0
python-dotenv>=1.0.0
tenacity>=8.2.0
plotly>=5.0.0
